+++ src/store/earning-management/moduleEarningManagementMutations.js
@@ -1,10 +1,15 @@
+import { rearrangeData, updateArrays } from '@/utils/utils';
+
 export default {
   GET_EARNINGS(state, { earnings, dates }) {
     state.earnings = earnings
+    state.groupearnings = rearrangeData(earnings)
     state.dates = dates
   },
-  CHANGE_EARNINGS(state, { changedatas }) {
+  CHANGE_EARNING(state, { changedata }) {
+    let changedatas = updateArrays (state.earnings, changedata)
     state.earnings = changedatas
+    state.groupearnings = rearrangeData(changedatas)
   },
   
   GET_YEARMONTHS(state, { yearmonths }) {
@@ -12,14 +17,8 @@ export default {
   },
 
   UPDATE_YEARMONTH(state, { updatedata }) {
-    let data = state.yearmonths;
-    data.forEach((item, index) => {
-      if (item._id === updatedata._id) {
-        data[index] = { ...item, ...updatedata };
-      }
-    });
-    data.push(updatedata);
-    state.yearmonths = data
+    let changedatas = updateArrays (state.yearmonths, updatedata)
+    state.yearmonths = changedatas
   },
   
   ADD_YEARMONTH(state, { adddata }) {
@@ -28,25 +27,4 @@ export default {
     state.yearmonths = data
   },
 
-  REMOVE_RECORD(state, itemId) {
-    const userIndex = state.users.findIndex((u) => u.userID === itemId)
-
-    state.users.splice(userIndex, 1)
-  },
-  UPDATE_USER(state, data) {
-    const userIndex = state.users.findIndex((u) => u.userID === data.user.userID)
-
-    if (data.path != '') state.users[userIndex].avatar = data.path
-    // if (data.user.userID == this.$store.state.AppActiveUser) localStorage.setItem('userInfo') = state.users[userIndex]
-  },
-  APPROVE_USER(state, data) {
-    const userIndex = state.users.findIndex((u) => u.userID === data)
-
-    state.users[userIndex].statue = true
-  },
-  REJECT_USER(state, data) {
-    const userIndex = state.users.findIndex((u) => u.userID === data)
-
-    state.users[userIndex].statue = false
-  }
 }
