+++ src/store/earning-management/moduleEarningManagementState.js
@@ -1,4 +1,5 @@
 export default {
+  groupearnings: [],
   earnings: [],
   dates: [],
   yearmonths:[],
