+++ src/views/main/earning/setting/Index.vue
@@ -30,13 +30,26 @@
             {{ data[indextr].end_date }}
           </vs-td>
           <vs-td>
-            <div class="grid-layout-container alignment-block">
-              <vs-row vs-type="flex" vs-w="12">
-                <vs-button color="primary" type="filled" @click="editPrompt(data[indextr])">Edit</vs-button>
-              </vs-row>
-            </div>
+
+          <div :style="{ direction: $vs.rtl ? 'rtl' : 'ltr' }">
+            <feather-icon
+              icon="EditIcon"
+              svgClasses="h-5 w-5 mr-4 text-primary cursor-pointer"
+              @click="editPrompt(data[indextr])"
+            />
+            <!-- <feather-icon
+              icon="Trash2Icon"
+              svgClasses="h-5 w-5 text-danger cursor-pointer"
+              @click="confirmDeleteRecord(data[indextr])"
+            /> -->
+          </div>
+
           </vs-td>
           <vs-prompt
+            title="Edit Year&Month"
+            accept-text= "Save"
+            cancel-text = "Cancel"
+            button-cancel = "border"
             @cancel="val = ''"
             @accept="editYearMonth"
             @close="close"
@@ -45,15 +58,32 @@
             <div class="flex flex-wrap">
               <div class="w-full bg-grid-color">
                 <label class="text-sm opacity-75">Year & Month</label>
-                <datepicker class="w-full" size="large" :minimumView="'month'" :maximumView="'year'" :format="yearmonthformat" v-model="edit_yearmonth"></datepicker>
+                <datepicker
+                  class="w-full"
+                  size="large"
+                  :minimumView="'month'"
+                  :maximumView="'year'"
+                  :format="yearmonthformat"
+                  v-model="edit_yearmonth"
+                ></datepicker>
               </div>
               <div class="w-full bg-grid-color">
                 <label class="text-sm opacity-75">Start Date</label>
-                <datepicker :format="start_dateformat" class="w-full" size="large" v-model="edit_start_date" />
+                <datepicker
+                  :format="start_dateformat"
+                  class="w-full"
+                  size="large"
+                  v-model="edit_start_date"
+                />
               </div>
               <div class="w-full bg-grid-color">
                 <label class="text-sm opacity-75">End Date</label>
-                <datepicker :format="end_dateformat" class="w-full" size="large" v-model="edit_end_date" />
+                <datepicker
+                  :format="end_dateformat"
+                  class="w-full"
+                  size="large"
+                  v-model="edit_end_date"
+                />
               </div>
             </div>
           </vs-prompt>
@@ -61,6 +91,10 @@
       </template>
     </vs-table>
     <vs-prompt
+      title="Add Year&Month"
+      accept-text= "Add"
+      cancel-text = "Cancel"
+      button-cancel = "border"
       @cancel="val = ''"
       @accept="addYearMonth"
       @close="close"
@@ -80,11 +114,21 @@
         </div>
         <div class="w-full bg-grid-color">
           <label class="text-sm opacity-75">Start Date</label>
-          <datepicker :format="start_dateformat" class="w-full" size="large" v-model="start_date" />
+          <datepicker
+            :format="start_dateformat"
+            class="w-full"
+            size="large"
+            v-model="start_date"
+          />
         </div>
         <div class="w-full bg-grid-color">
           <label class="text-sm opacity-75">End Date</label>
-          <datepicker :format="end_dateformat" class="w-full" size="large" v-model="end_date" />
+          <datepicker
+            :format="end_dateformat"
+            class="w-full"
+            size="large"
+            v-model="end_date"
+          />
         </div>
       </div>
     </vs-prompt>
@@ -103,17 +147,19 @@ export default {
       yearmonthformat: "yyyy-MM",
       start_dateformat: "MM-dd",
       end_dateformat: "MM-dd",
-      start_date: new Date,
-      end_date: new Date,
+      start_date: new Date(),
+      end_date: new Date(),
       yearmonth: new Date(),
-      
+
       edit_id: null,
       edit_start_date: null,
       edit_end_date: null,
       edit_yearmonth: null,
-      
+
       activePrompt: false,
       isEditPrompt: false,
+      
+      delete_id: null,
     };
   },
   computed: {
@@ -132,30 +178,30 @@ export default {
       };
       this.$store.dispatch("earningManagement/addYearMonth", payload);
     },
-    editPrompt (data) {
+    editPrompt(data) {
+      const [year, month] = data.yearmonth.split("-");
+      const [start_month, start_day] = data.start_date.split("-");
+      const [end_month, end_day] = data.end_date.split("-");
 
-      const [year,month] = data.yearmonth.split('-');
-      const [start_month, start_day] = data.start_date.split('-');
-      const [end_month, end_day] = data.end_date.split('-');
-
-      this.edit_id = data._id
-      this.edit_yearmonth = data.yearmonth
-      this.edit_start_date = new Date(`${year}-${start_month}-${start_day}`)
-      this.edit_end_date = new Date(`${year}-${end_month}-${end_day}`)
-      this.isEditPrompt=true;
+      this.edit_id = data._id;
+      this.edit_yearmonth = data.yearmonth;
+      this.edit_start_date = new Date(`${year}-${start_month}-${start_day}`);
+      this.edit_end_date = new Date(`${year}-${end_month}-${end_day}`);
+      this.isEditPrompt = true;
     },
     async editYearMonth() {
-
       const payload = {
-        _id:this.edit_id,
+        _id: this.edit_id,
         year: new Date(this.edit_yearmonth).getFullYear(),
         month: new Date(this.edit_yearmonth).getMonth() + 1,
-        start_date: `${this.edit_start_date.getMonth() + 1}-${this.edit_start_date.getDate()}`,
+        start_date: `${
+          this.edit_start_date.getMonth() + 1
+        }-${this.edit_start_date.getDate()}`,
         end_date: `${this.edit_end_date.getMonth() + 1}-${this.edit_end_date.getDate()}`,
         notify: this.$vs.notify,
       };
-      await this.$store.dispatch("earningManagement/updateYearMonth", payload,);
-      // await this.$store.dispatch("earningManagement/updateYearMonth", payload, this.updateData(this.yearmonths, payload));
+      await this.$store.dispatch("earningManagement/updateYearMonth", payload);
+      this.$router.go(0);
     },
 
     updateData(datas, inputData) {
@@ -166,6 +212,26 @@ export default {
         return item;
       });
     },
+
+    confirmDeleteRecord(data) {
+      this.delete_id = data._id
+      this.$vs.dialog({
+        type: "confirm",
+        color: "danger",
+        title: "Confirm Delete",
+        text: `Are you really delete "${data.yearmonth}"`,
+        accept: this.deleteRecord,
+        acceptText: "Delete",
+      });
+    },
+    async deleteRecord() {
+      const payload = {
+        _id: this.delete_id,
+      };
+      await this.$store.dispatch("earningManagement/removeRecord", payload);
+      this.$vs.$forceUpdate();
+    },
+
     close() {
       this.$vs.notify({
         color: "danger",
@@ -174,8 +240,7 @@ export default {
       });
     },
   },
-  mounted() {
-  },
+  mounted() {},
   beforeCreate() {
     if (!moduleEarningManagement.isRegistered) {
       this.$store.registerModule("earningManagement", moduleEarningManagement);
@@ -185,8 +250,9 @@ export default {
       year: new Date().getFullYear(),
       notify: this.$vs.notify,
     };
-    this.$store.dispatch("earningManagement/fetchYearMonths", payload)
-    .then((res) => {})
+    this.$store
+      .dispatch("earningManagement/fetchYearMonths", payload)
+      .then((res) => {})
       .catch((err) => {
         this.user_not_found = true;
         return;
